Simplify Product Form (#253)
