+++ ui/src/Products/ProductForm.tsx
@@ -15,42 +15,53 @@
  * limitations under the License.
  */
 
-import React, {ChangeEvent, CSSProperties, useEffect, useState} from 'react';
+import React, {ChangeEvent, CSSProperties, useState} from 'react';
+import {GlobalStateProps} from '../Redux/Reducers';
+import {Dispatch} from 'redux';
+import {connect} from 'react-redux';
+import {closeModalAction, setAllGroupedTagFilterOptions} from '../Redux/Actions';
+import {JSX} from '@babel/types';
+import {StylesConfig} from 'react-select';
+import moment from 'moment';
 import ProductClient from './ProductClient';
+import {emptyProduct, Product} from './Product';
+import ConfirmationModal, {ConfirmationModalProps} from '../Modal/ConfirmationModal';
+import {reactSelectStyles} from '../ReusableComponents/ReactSelectStyles';
+import {ProductTag} from '../ProductTag/ProductTag';
+import {FilterOption} from '../CommonTypes/Option';
+import {AllGroupedTagFilterOptions} from '../ReusableComponents/ProductFilter';
+import {Trait} from '../Traits/Trait';
+import {Space} from '../SpaceDashboard/Space';
+import ProductFormLocationField from './ProductFormLocationField';
+import ProductFormProductTagsField from './ProductFormProductTagsField';
+import ProductFormStartDateField from './ProductFormStartDateField';
+import ProductFormEndDateField from './ProductFormEndDateField';
 
+import 'react-datepicker/dist/react-datepicker.css';
 import '../Modal/Form.scss';
 import './ProductForm.scss';
 
-import LocationClient from '../Locations/LocationClient';
-import {closeModalAction, setAllGroupedTagFilterOptions} from '../Redux/Actions';
-import {connect} from 'react-redux';
-import DatePicker from 'react-datepicker';
-import MaskedInput from 'react-text-mask';
-import 'react-datepicker/dist/react-datepicker.css';
-import {JSX} from '@babel/types';
-import {emptyProduct, Product} from './Product';
-import ConfirmationModal, {ConfirmationModalProps} from '../Modal/ConfirmationModal';
-import {CreateNewText, CustomIndicator, CustomOption, reactSelectStyles} from '../ReusableComponents/ReactSelectStyles';
-import Creatable from 'react-select/creatable';
-import {ProductTag} from '../ProductTag/ProductTag';
-import ProductTagClient from '../ProductTag/ProductTagClient';
-import {AxiosResponse} from 'axios';
-import {SpaceLocation} from '../Locations/SpaceLocation';
-import {FilterOption, Option} from '../CommonTypes/Option';
-import {TraitAddRequest} from '../Traits/TraitAddRequest';
-import {GlobalStateProps} from '../Redux/Reducers';
-import {AllGroupedTagFilterOptions} from '../ReusableComponents/ProductFilter';
-import {Trait} from '../Traits/Trait';
-import {StylesConfig} from 'react-select';
-import {Dispatch} from 'redux';
-import moment from 'moment';
-import {Space} from '../SpaceDashboard/Space';
+export const customStyles: StylesConfig = {
+    ...reactSelectStyles,
+    valueContainer: (provided: CSSProperties) => ({
+        ...provided,
+        padding: '0px 3px',
+    }),
+    multiValue: (provided: CSSProperties) => ({
+        ...provided,
+        alignItems: 'center',
+        backgroundColor: '#F2E7F3',
+        fontFamily: 'Helvetica, sans-serif',
+        borderRadius: '6px',
+        height: '22px',
+        marginRight: '3px',
+    }),
+};
 
 interface ProductFormProps {
     editing: boolean;
     product?: Product;
     spaceId: number;
-    currentSpace: Space;
     viewingDate: string;
     allGroupedTagFilterOptions: Array<AllGroupedTagFilterOptions>;
     setAllGroupedTagFilterOptions(groupedTagFilterOptions: Array<AllGroupedTagFilterOptions>): void;
@@ -61,35 +72,12 @@ function ProductForm({
     editing,
     product,
     spaceId,
-    currentSpace,
     viewingDate,
     allGroupedTagFilterOptions,
     setAllGroupedTagFilterOptions,
     closeModal,
 }: ProductFormProps): JSX.Element {
-
-    const customStyles: StylesConfig = {
-        ...reactSelectStyles,
-        valueContainer: (provided: CSSProperties) => ({
-            ...provided,
-            padding: '0px 3px',
-        }),
-        multiValue: (provided: CSSProperties) => ({
-            ...provided,
-            alignItems: 'center',
-            backgroundColor: '#F2E7F3',
-            fontFamily: 'Helvetica, sans-serif',
-            borderRadius: '6px',
-            height: '22px',
-            marginRight: '3px',
-        }),
-    };
-
     const [currentProduct, setCurrentProduct] = useState<Product>(initializeProduct());
-
-    const [availableLocations, setAvailableLocations] = useState<SpaceLocation[]>([]);
-    const [availableProductTags, setAvailableProductTags] = useState<Array<ProductTag>>([]);
-
     const [selectedProductTags, setSelectedProductTags] = useState<Array<ProductTag>>([]);
 
     const [isLoading, setIsLoading] = useState<boolean>(false);
@@ -97,20 +85,6 @@ function ProductForm({
 
     const [duplicateProductNameWarning, setDuplicateProductNameWarning] = useState<boolean>(false);
     const [notesFieldLength, setNotesFieldLength] = useState<number>(product && product.notes ? product.notes.length : 0);
-    const [typedInProductTag, setTypedInProductTag] = useState<string>('');
-
-    const [typedInLocation, setTypedInLocation] = useState<string>('');
-
-    const [startDate, setStartDate] = useState<Date>(currentProduct.startDate ? moment(currentProduct.startDate).toDate() : moment(viewingDate).toDate());
-    const [endDate, setEndDate] = useState<Date | null>(currentProduct.endDate ? moment(currentProduct.endDate).toDate() : null);
-
-    useEffect(() => {
-        LocationClient.get(currentSpace.name).then(result => {setAvailableLocations(result.data);});
-        ProductTagClient.get(currentSpace.name).then(result => setAvailableProductTags(result.data));
-
-        setSelectedProductTags(currentProduct.productTags);
-
-    }, []);
 
     function initializeProduct(): Product {
         if (product == null) {
@@ -120,7 +94,6 @@ function ProductForm({
     }
 
     // Put the selected product tags on the updated product and use the 'editProduct' endpoint
-
     function handleSubmit(): void {
         currentProduct.productTags = selectedProductTags;
 
@@ -174,89 +147,6 @@ function ProductForm({
         setCurrentProduct(updatedProduct);
     }
 
-    function updateSelectedProductTags(productTags: Array<ProductTag>): void {
-        if (productTags.length > 0) {
-            setSelectedProductTags([...productTags]);
-        } else {
-            setSelectedProductTags([]);
-        }
-    }
-
-    function optionToProductTag(options: Array<Option>): Array<ProductTag> {
-        if (options) {
-            return options.map(option => {
-                return {
-                    id: Number.parseInt(option.value, 10),
-                    name: option.label,
-                    spaceId,
-                };
-            });
-        } else {
-            return [];
-        }
-    }
-
-    function optionToSpaceLocation(option: Option): SpaceLocation {
-        return {
-            id: Number.parseInt(option.value.split('_')[0], 10),
-            name: option.label,
-            spaceId,
-        };
-    }
-
-    function createLocationOption(location: SpaceLocation): Option {
-        return {
-            label: location.name,
-            value: location.id!.toString(),
-        };
-    }
-
-    function createTagOption(label: string, id: number): Option {
-        return {
-            label: label,
-            value: id.toString() + '_' + label,
-        };
-    }
-
-    function locationOptions(): Option[] {
-        return availableLocations.map(location => createLocationOption(location));
-    }
-
-    function handleCreateProductTag(inputValue: string): void {
-        setIsLoading(true);
-        const productTag: TraitAddRequest = {
-            name: inputValue,
-        };
-        ProductTagClient.add(productTag, currentSpace.name).then((response: AxiosResponse) => {
-            const newProductTag: ProductTag = response.data;
-            setAvailableProductTags(productTags => [...productTags, {
-                id: newProductTag.id,
-                name: newProductTag.name,
-            }] as Array<ProductTag>);
-            addGroupedTagFilterOptions(1, newProductTag as Trait);
-            updateSelectedProductTags([...selectedProductTags, newProductTag]);
-            setIsLoading(false);
-        });
-    }
-
-    function handleCreateLocationTag(inputValue: string): void {
-        setIsLoading(true);
-
-        const location: TraitAddRequest = {
-            name: inputValue,
-        };
-        LocationClient.add(location, currentSpace.name).then((result: AxiosResponse) => {
-            const newLocation: SpaceLocation = result.data;
-            setAvailableLocations([...availableLocations, newLocation]);
-            addGroupedTagFilterOptions(0, newLocation as Trait);
-            setCurrentProduct({
-                ...currentProduct,
-                spaceLocation: newLocation,
-            });
-            setIsLoading(false);
-        });
-    }
-
     function addGroupedTagFilterOptions(tagFilterIndex: number, trait: Trait): void {
         const addedFilterOption: FilterOption = {
             label: trait.name,
@@ -274,14 +164,6 @@ function ProductForm({
         let groupedTagFilterOptions: Array<AllGroupedTagFilterOptions> = [...allGroupedTagFilterOptions];
         groupedTagFilterOptions[tagFilterIndex] = updatedTagFilterOptions;
         setAllGroupedTagFilterOptions(groupedTagFilterOptions);
-    }
-
-    function locationOptionValue(): Option | undefined {
-        if (currentProduct.spaceLocation && currentProduct.spaceLocation.name !== '') {
-            return createLocationOption(currentProduct.spaceLocation);
-        }
-
-        return undefined;
     }
 
     function notesChanged(e: ChangeEvent<HTMLTextAreaElement>): void {
@@ -306,92 +188,27 @@ function ProductForm({
                     {duplicateProductNameWarning &&
                     <span className="personNameWarning">A product with this name already exists. Please enter a different name.</span>}
                 </div>
-                <div className="formItem">
-                    <label className="formItemLabel" htmlFor="location">Location</label>
-                    <Creatable
-                        name="location"
-                        inputId="location"
-                        onInputChange={(e: string): void => setTypedInLocation(e)}
-                        onChange={(e): void  => updateProductField('spaceLocation', optionToSpaceLocation(e as Option))}
-                        isLoading={isLoading}
-                        isDisabled={isLoading}
-                        onCreateOption={handleCreateLocationTag}
-                        options={locationOptions()}
-                        styles={customStyles}
-                        components={{DropdownIndicator: CustomIndicator, Option: CustomOption}}
-                        formatCreateLabel={(): JSX.Element => CreateNewText(`Create "${typedInLocation}"`)}
-                        placeholder="Select or create location"
-                        hideSelectedOptions={true}
-                        isClearable={false}
-                        value={locationOptionValue()}
-                    />
-                </div>
-                <div className="formItem">
-                    <label className="formItemLabel" htmlFor="productTags">Product Tags</label>
-                    <Creatable
-                        isMulti={true}
-                        name="productTags"
-                        inputId="productTags"
-                        onInputChange={(e: string): void => setTypedInProductTag(e)}
-                        onChange={(e): void => updateSelectedProductTags(optionToProductTag(e as Option[]))}
-                        isLoading={isLoading}
-                        isDisabled={isLoading}
-                        onCreateOption={handleCreateProductTag}
-                        options={availableProductTags.map((productTag: ProductTag) => createTagOption(productTag.name, productTag.id))}
-                        styles={customStyles}
-                        value={selectedProductTags.map(productTag => createTagOption(productTag.name, productTag.id))}
-                        components={{DropdownIndicator: CustomIndicator, Option: CustomOption}}
-                        formatCreateLabel={(): JSX.Element => CreateNewText(`Create "${typedInProductTag}"`)}
-                        placeholder="Select or create product tags"
-                        hideSelectedOptions={true}
-                        isClearable={false}
-                    />
-                </div>
-                <div className="formItem" data-testid="productFormStartDateField">
-                    <label className="formItemLabel" htmlFor="start">Start Date</label>
-                    <DatePicker
-                        className="formInput formTextInput"
-                        name="start"
-                        id="start"
-                        selected={startDate}
-                        onChange={date => {
-                            if (date) {
-                                setStartDate(date);
-                                updateProductField('startDate', moment(date).format('YYYY-MM-DD'));
-                            } else {
-                                setStartDate(moment(currentProduct.startDate).toDate());
-                                updateProductField('startDate', moment(currentProduct.startDate).format('YYYY-MM-DD'));
-                            }
-                        }}
-                        customInput={
-                            <MaskedInput
-                                mask={[/\d/, /\d/, '/', /\d/, /\d/, '/', /\d/, /\d/, /\d/, /\d/]}
-                            />
-                        }
-                    />
-                    <i className="far fa-calendar-alt calendar-icon" />
-                </div>
-                <div className="formItem" data-testid="productFormNextPhaseDateField">
-                    <label className="formItemLabel" htmlFor="end">End Date</label>
-                    <DatePicker
-                        className="formInput formTextInput"
-                        name="end"
-                        id="end"
-                        selected={endDate}
-                        onChange={date => {
-                            setEndDate(date ? date : null);
-                            updateProductField('endDate', date ? moment(date).format('YYYY-MM-DD') : '');
-                        }}
-                        customInput={
-                            <MaskedInput
-                                mask={[/\d/, /\d/, '/', /\d/, /\d/, '/', /\d/, /\d/, /\d/, /\d/]}
-                            />
-                        }
-                        isClearable
-                        placeholderText="MM/DD/YYYY"
-                    />
-                    {!endDate && <i className="far fa-calendar-alt calendar-icon" />}
-                </div>
+                <ProductFormLocationField
+                    spaceId={spaceId}
+                    currentProductState={{ currentProduct, setCurrentProduct }}
+                    loadingState={{ isLoading, setIsLoading }}
+                    addGroupedTagFilterOptions={addGroupedTagFilterOptions}
+                />
+                <ProductFormProductTagsField
+                    spaceId={spaceId}
+                    currentProductState={{ currentProduct }}
+                    loadingState={{ isLoading, setIsLoading }}
+                    selectedProductTagsState={{ selectedProductTags, setSelectedProductTags }}
+                    addGroupedTagFilterOptions={addGroupedTagFilterOptions}
+                />
+                <ProductFormStartDateField
+                    currentProduct={currentProduct}
+                    updateProductField={updateProductField}
+                />
+                <ProductFormEndDateField
+                    currentProduct={currentProduct}
+                    updateProductField={updateProductField}
+                />
                 <div className="formItem">
                     <label className="formItemLabel" htmlFor="notes">Notes</label>
                     <textarea
@@ -433,7 +250,6 @@ function ProductForm({
     );
 }
 const mapStateToProps = (state: GlobalStateProps) => ({
-    currentSpace: state.currentSpace,
     viewingDate: moment(state.viewingDate).format('YYYY-MM-DD'),
     allGroupedTagFilterOptions: state.allGroupedTagFilterOptions,
 });
