+++ ui/src/Products/ProductFormProductTagsField.tsx
@@ -1,0 +1,128 @@
+import Creatable from 'react-select/creatable';
+import {Option} from '../CommonTypes/Option';
+import {ProductTag} from '../ProductTag/ProductTag';
+import {CreateNewText, CustomIndicator, CustomOption} from '../ReusableComponents/ReactSelectStyles';
+import {JSX} from '@babel/types';
+import React, {useEffect, useState} from 'react';
+import {customStyles} from './ProductForm';
+import {Product} from './Product';
+import {Trait} from '../Traits/Trait';
+import {Space} from '../SpaceDashboard/Space';
+import {TraitAddRequest} from '../Traits/TraitAddRequest';
+import ProductTagClient from '../ProductTag/ProductTagClient';
+import {AxiosResponse} from 'axios';
+import {GlobalStateProps} from '../Redux/Reducers';
+import {connect} from 'react-redux';
+
+interface Props {
+    spaceId: number;
+    loadingState: { isLoading: boolean; setIsLoading: (isLoading: boolean) => void };
+    currentProductState: { currentProduct: Product };
+    selectedProductTagsState: {
+        selectedProductTags: Array<ProductTag>;
+        setSelectedProductTags: (productTags: Array<ProductTag>) => void;
+    };
+    addGroupedTagFilterOptions: (tagFilterIndex: number, trait: Trait) => void;
+    currentSpace: Space;
+}
+
+function ProductFormProductTagsField({
+    spaceId,
+    loadingState: {
+        isLoading,
+        setIsLoading,
+    },
+    currentProductState: {
+        currentProduct,
+    },
+    selectedProductTagsState: {
+        selectedProductTags,
+        setSelectedProductTags,
+    },
+    currentSpace,
+    addGroupedTagFilterOptions,
+}: Props): JSX.Element {
+    const [typedInProductTag, setTypedInProductTag] = useState<string>('');
+    const [availableProductTags, setAvailableProductTags] = useState<Array<ProductTag>>([]);
+
+    useEffect(() => {
+        ProductTagClient.get(currentSpace.name).then(result => setAvailableProductTags(result.data));
+
+        setSelectedProductTags(currentProduct.productTags);
+    }, []);
+
+    function createTagOption(label: string, id: number): Option {
+        return {
+            label: label,
+            value: id.toString() + '_' + label,
+        };
+    }
+
+    function optionToProductTag(options: Array<Option>): Array<ProductTag> {
+        if (options) {
+            return options.map(option => {
+                return {
+                    id: Number.parseInt(option.value, 10),
+                    name: option.label,
+                    spaceId,
+                };
+            });
+        }
+        return [];
+    }
+    
+    function handleCreateProductTag(inputValue: string): void {
+        setIsLoading(true);
+        const productTag: TraitAddRequest = { name: inputValue };
+
+        ProductTagClient.add(productTag, currentSpace.name)
+            .then((response: AxiosResponse) => {
+                const newProductTag: ProductTag = response.data;
+                setAvailableProductTags(productTags => [...productTags, {
+                    id: newProductTag.id,
+                    name: newProductTag.name,
+                }] as Array<ProductTag>);
+                addGroupedTagFilterOptions(1, newProductTag as Trait);
+                updateSelectedProductTags([...selectedProductTags, newProductTag]);
+                setIsLoading(false);
+            });
+    }
+
+    function updateSelectedProductTags(productTags: Array<ProductTag>): void {
+        if (productTags.length > 0) {
+            setSelectedProductTags([...productTags]);
+        } else {
+            setSelectedProductTags([]);
+        }
+    }
+    
+    return (
+        <div className="formItem">
+            <label className="formItemLabel" htmlFor="productTags">Product Tags</label>
+            <Creatable
+                isMulti={true}
+                name="productTags"
+                inputId="productTags"
+                onInputChange={(e: string): void => setTypedInProductTag(e)}
+                onChange={(option): void => updateSelectedProductTags(optionToProductTag(option as Option[]))}
+                isLoading={isLoading}
+                isDisabled={isLoading}
+                onCreateOption={handleCreateProductTag}
+                options={availableProductTags.map((productTag: ProductTag) => createTagOption(productTag.name, productTag.id))}
+                styles={customStyles}
+                value={selectedProductTags.map(productTag => createTagOption(productTag.name, productTag.id))}
+                components={{DropdownIndicator: CustomIndicator, Option: CustomOption}}
+                formatCreateLabel={(): JSX.Element => CreateNewText(`Create "${typedInProductTag}"`)}
+                placeholder="Select or create product tags"
+                hideSelectedOptions={true}
+                isClearable={false}
+            />
+        </div>
+    );
+}
+
+const mapStateToProps = (state: GlobalStateProps) => ({
+    currentSpace: state.currentSpace,
+});
+
+export default connect(mapStateToProps)(ProductFormProductTagsField);
